Feature/kr 65

+++ domain-application-form/src/main/java/co/dalicious/domain/application_form/validator/ApplicationFormValidator.java
@@ -17,7 +17,7 @@ public class ApplicationFormValidator {
     private final ApartmentApplicationFormRepository apartmentApplicationFormRepository;
     private final CorporationApplicationFormRepository corporationApplicationFormRepository;
 
-    public ApartmentApplicationForm isValidApartmentApplicationForm(BigInteger userId, Long applicationFormId) {
+    public ApartmentApplicationForm isValidApartmentApplicationForm(BigInteger userId, BigInteger applicationFormId) {
         // 신청서가 존재하는지 확인
         ApartmentApplicationForm apartmentApplicationForm = apartmentApplicationFormRepository.findById(applicationFormId)
                 .orElseThrow(() -> new ApiException(ExceptionEnum.APPLICATION_FORM_NOT_FOUND));
@@ -28,7 +28,7 @@ public class ApplicationFormValidator {
         return apartmentApplicationForm;
     }
 
-    public CorporationApplicationForm isValidCorporationApplicationForm(BigInteger userId, Long applicationFormId) {
+    public CorporationApplicationForm isValidCorporationApplicationForm(BigInteger userId, BigInteger applicationFormId) {
         // 신청서가 존재하는지 확인
         CorporationApplicationForm corporationApplicationForm = corporationApplicationFormRepository.findById(applicationFormId)
                 .orElseThrow(() -> new ApiException(ExceptionEnum.APPLICATION_FORM_NOT_FOUND));
